Some Minor Changes wrt prev sarv commit

+++ backend/api/models.py
@@ -5,8 +5,8 @@ from typing import DefaultDict, List, Dict
 # Model 1: For Coordinate data
 class coordinates(BaseModel): 
 
-    coord_x: int = Field(..., examples=1)
-    coord_y: int = Field(..., examples=1)
-    coord_z: int = Field(..., examples=1)
+    coord_x: float = Field(..., examples=1)
+    coord_y: float = Field(..., examples=1)
+    coord_z: float = Field(..., examples=1)
